requirements.txt
django
django-sequences